bug 3059 - Fixed taglib extension to not crash on files shorter than 1024 bytes.
Replayed for 1.6.1 from 10079
Replayed for 1.5 from 4281


git-svn-id: http://publicsvn.songbirdnest.com/vendor/trunk@10851 37441250-4157-47a9-a2c9-f9fe3212d2d3

+++ taglib/taglib/mpeg/mpegfile.cpp
@@ -333,7 +333,7 @@ long MPEG::File::nextFrameOffset(long position)
     if(foundLastSyncPattern && secondSynchByte(buffer[0]))
       return position - 1;
 
-    for(uint i = 0; i < buffer.size() - 1; i++) {
+    for(int i = 0; i < ((int) buffer.size()) - 1; i++) {
       if(uchar(buffer[i]) == 0xff && secondSynchByte(buffer[i + 1]))
         return position + i;
     }
